# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Obras",
    "description": "API de Obras",
    "contact": {
      "name": "Tecnología y Negocios Pelom",
      "url": "https://www.pelom.cl",
      "email": "ti.pelom@pelom.cl"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/reportes/v1/alljornada": {
      "get": {
        "tags": [
          "SAE - Backoffice - Reportes"
        ],
        "description": "Devuelve todas las jornadas",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/reportes/v1/alleventos": {
      "get": {
        "tags": [
          "SAE - Backoffice - Reportes"
        ],
        "description": "Devuelve todos los eventos",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/reportes/v1/allestadosresultado": {
      "get": {
        "tags": [
          "SAE - Backoffice - Reportes"
        ],
        "description": "Devuelve todos los estados de resultado",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/reportes/v1/creaestadoresultado": {
      "post": {
        "tags": [
          "SAE - Backoffice - Reportes"
        ],
        "description": "Crea un estado resultado",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "eventos_relacionados": {
                  "example": "any"
                },
                "id_usuario": {
                  "example": "any"
                },
                "zona": {
                  "example": "any"
                },
                "paquete": {
                  "example": "any"
                },
                "mes": {
                  "example": "any"
                },
                "fecha_inicio": {
                  "example": "any"
                },
                "fecha_final": {
                  "example": "any"
                },
                "nombre_doc": {
                  "example": "any"
                },
                "url_doc": {
                  "example": "any"
                },
                "fecha_creacion": {
                  "example": "any"
                },
                "fecha_modificacion": {
                  "example": "any"
                },
                "estado": {
                  "example": "any"
                },
                "id_cliente": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/reportes/v1/resumeneventos": {
      "get": {
        "tags": [
          "SAE - Backoffice - Reportes"
        ],
        "description": "Devuelve el resumen de eventos por paquete y rango de fechas",
        "parameters": [
          {
            "name": "id_paquete",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fecha_inicial",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fecha_final",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/reportes/v1/resumenturnos": {
      "get": {
        "tags": [
          "SAE - Backoffice - Reportes"
        ],
        "description": "Devuelve el resumen de turnos por paquete y rango de fechas",
        "parameters": [
          {
            "name": "id_paquete",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fecha_inicial",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fecha_final",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}